--- Classes/Pagerange/Markdown/ParserNotFoundException.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Pagerange\Markdown;
-
-class ParserNotFoundException extends \Exception
-{
-	
-}
