+++ Metaparsedown/Pagerange/Markdown/ParserNotFoundException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Metaparsedown\Pagerange\Markdown;
+
+class ParserNotFoundException extends \Exception
+{
+	
+}
